# Simple manual proof system benchmark