this adds a small benchmark script with the following results:

without pickles:
```
circuit size (without pickles overhead) 49152
compile 1 (includes srs creation)... 54.515 sec
compile 2... 10.954 sec
prove... 28.018 sec
verify... 0.660 sec
```

with pickles:
```
circuit size (without pickles overhead) 49152
compile 1 (includes srs creation)... 68.329 sec
compile 2... 19.319 sec
prove... 41.963 sec
verify... 1.074 sec
```

run the script with:
```bash
./run scr/examples/benchmarks/mul.ts --bundle
```

EDIT: this now also adds a way to serve a TS script for running in the browser. added a version of the benchmark that doesn't use nodejs builtins to be able to use it:

```
./run-in-browser.js src/examples/benchmarks/mul-web.ts
```

this can be helpful to inspect the benchmark in chrome devtools: